cloud_run_jobs/moi_codebase/requirements.txt
Shapely==1.8.5.post1
pyshp==2.3.1
geopandas==0.12.1
requests==2.28.1
google-cloud-bigquery==3.4.0
